refactor(core): move withIfMatch to src/core, strengthen error-message test (C6 #304)
withIfMatch has no HTTP/Hono dependencies — it's a pure async retry
combinator over CasMutationResult. Moving it to src/core eliminates
the only core→server runtime import edge in the codebase.

Also strengthens the retry-exhausted test to verify the documented
"last current RV=X" substring in the error message — protects the
contract from drift.
